Split large orchestration and runtime modules
