+++ src/backend/services/session/service/acp/acp-runtime-manager.ts
@@ -10,10 +10,6 @@ import {
   ndJsonStream,
   PROTOCOL_VERSION,
   type SessionConfigOption,
-  type SessionConfigSelectGroup,
-  type SessionConfigSelectOption,
-  type SessionModelState,
-  type SessionModeState,
 } from '@agentclientprotocol/sdk';
 import pLimit from 'p-limit';
 import { createLogger, getCurrentProcessEnv } from '@/backend/services/logger.service';
@@ -21,6 +17,8 @@ import { AcpClientHandler, type AutoApprovePolicy } from './acp-client-handler';
 import type { AcpPermissionBridge } from './acp-permission-bridge';
 import { AcpProcessHandle } from './acp-process-handle';
 import type { AcpRuntimeEvent } from './acp-runtime-events';
+import { requireSessionConfigOptions } from './acp-session-config-options';
+import { createNormalizedAcpReadableStream, normalizeUnknownError } from './acp-stream-normalizer';
 import type { AcpClientOptions, PermissionPreset } from './types';
 
 const logger = createLogger('acp-runtime-manager');
@@ -91,103 +89,6 @@ type AcpErrorLogDetails = {
   code?: number | string;
   data?: unknown;
 };
-
-function normalizeUnknownError(error: unknown): Error {
-  return error instanceof Error ? error : new Error(String(error));
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-/**
- * Workaround for claude-agent-acp bug where the Read tool sends `line: [start, end]`
- * (an array) instead of `line: number` in session/update locations.
- * Normalizes array line values to the first element before SDK Zod validation.
- */
-function normalizeSessionUpdateMessage(message: unknown): unknown {
-  if (
-    !isRecord(message) ||
-    message.method !== 'session/update' ||
-    !isRecord(message.params) ||
-    !isRecord(message.params.update) ||
-    !Array.isArray(message.params.update.locations)
-  ) {
-    return message;
-  }
-
-  const update = message.params.update;
-  const locations = update.locations as unknown[];
-  const needsNormalization = locations.some(
-    (loc) =>
-      isRecord(loc) && loc.line !== undefined && loc.line !== null && typeof loc.line !== 'number'
-  );
-  if (!needsNormalization) {
-    return message;
-  }
-
-  return {
-    ...message,
-    params: {
-      ...message.params,
-      update: {
-        ...update,
-        locations: locations.map((loc) => {
-          if (!isRecord(loc) || typeof loc.line === 'number' || loc.line == null) {
-            return loc;
-          }
-          return {
-            ...loc,
-            line:
-              Array.isArray(loc.line) && loc.line.length > 0 ? (loc.line[0] as number) : undefined,
-          };
-        }),
-      },
-    },
-  };
-}
-
-function createNormalizedAcpReadableStream<T>(readable: ReadableStream<T>): ReadableStream<T> {
-  if (typeof readable.getReader !== 'function') {
-    return readable;
-  }
-
-  let reader: ReadableStreamDefaultReader<T> | null = null;
-  return new ReadableStream<T>({
-    start() {
-      reader = readable.getReader();
-    },
-    async pull(controller) {
-      if (!reader) {
-        reader = readable.getReader();
-      }
-      const currentReader = reader;
-
-      try {
-        const { done, value } = await currentReader.read();
-        if (done) {
-          controller.close();
-          currentReader.releaseLock();
-          reader = null;
-          return;
-        }
-        controller.enqueue(normalizeSessionUpdateMessage(value) as T);
-      } catch (error) {
-        currentReader.releaseLock();
-        controller.error(error);
-        reader = null;
-      }
-    },
-    async cancel(reason) {
-      try {
-        await reader?.cancel(reason);
-      } finally {
-        reader?.releaseLock();
-        reader = null;
-      }
-    },
-  });
-}
 
 function hasUsableWorkingDir(workingDir: string | null | undefined): boolean {
   return typeof workingDir === 'string' && workingDir.trim().length > 0;
@@ -382,223 +283,6 @@ function getAcpErrorLogDetails(error: unknown): AcpErrorLogDetails {
 function isMethodNotFoundError(error: unknown): boolean {
   const details = getAcpErrorLogDetails(error);
   return details.code === -32_601 || details.message.includes('Method not found');
-}
-
-const REQUIRED_CONFIG_CATEGORIES = ['model', 'mode'] as const;
-
-function isNonEmptyString(value: unknown): value is string {
-  return typeof value === 'string' && value.length > 0;
-}
-
-function resolveModelFamilyName(description: string | null | undefined): string | null {
-  if (!isNonEmptyString(description)) {
-    return null;
-  }
-  const familyName = description.split('·')[0]?.trim();
-  return familyName && familyName.length > 0 ? familyName : null;
-}
-
-function normalizeClaudeModelOption(option: SessionConfigSelectOption): SessionConfigSelectOption {
-  if (option.value !== 'default') {
-    return option;
-  }
-  const modelFamilyName = resolveModelFamilyName(option.description);
-  if (!modelFamilyName) {
-    return option;
-  }
-  return {
-    ...option,
-    name: modelFamilyName,
-  };
-}
-
-function isOptionGroup(
-  option: SessionConfigSelectOption | SessionConfigSelectGroup
-): option is SessionConfigSelectGroup {
-  return 'group' in option;
-}
-
-function isOptionGroupArray(
-  options: SessionConfigOption['options']
-): options is SessionConfigSelectGroup[] {
-  const [firstOption] = options;
-  return firstOption ? isOptionGroup(firstOption) : false;
-}
-
-function normalizeClaudeConfigOptions(configOptions: SessionConfigOption[]): SessionConfigOption[] {
-  return configOptions.map((configOption) => {
-    if (configOption.category !== 'model') {
-      return configOption;
-    }
-
-    if (isOptionGroupArray(configOption.options)) {
-      const normalizedGroups = configOption.options.map((group) => ({
-        ...group,
-        options: group.options.map(normalizeClaudeModelOption),
-      }));
-
-      return {
-        ...configOption,
-        options: normalizedGroups,
-      };
-    }
-
-    const normalizedOptions = configOption.options.map(normalizeClaudeModelOption);
-
-    return {
-      ...configOption,
-      options: normalizedOptions,
-    };
-  });
-}
-
-function normalizeSessionConfigOptions(
-  provider: string,
-  configOptions: SessionConfigOption[]
-): SessionConfigOption[] {
-  const providerNormalized =
-    provider === 'CLAUDE' ? normalizeClaudeConfigOptions(configOptions) : configOptions;
-  return providerNormalized.map((configOption) => {
-    if (configOption.category || (configOption.id !== 'model' && configOption.id !== 'mode')) {
-      return configOption;
-    }
-    return {
-      ...configOption,
-      category: configOption.id,
-    };
-  });
-}
-
-type SessionResultWithFallbackState = {
-  configOptions?: SessionConfigOption[] | null;
-  models?: SessionModelState | null;
-  modes?: SessionModeState | null;
-};
-
-function resolveModelConfigOption(
-  models: SessionModelState | null | undefined
-): SessionConfigOption | null {
-  if (!(models && Array.isArray(models.availableModels))) {
-    return null;
-  }
-
-  const options = models.availableModels
-    .filter((entry) => isNonEmptyString(entry.modelId))
-    .map((entry) => ({
-      value: entry.modelId,
-      name: isNonEmptyString(entry.name) ? entry.name : entry.modelId,
-      ...(isNonEmptyString(entry.description) ? { description: entry.description } : {}),
-    }));
-  if (options.length === 0) {
-    return null;
-  }
-
-  const preferredValue = isNonEmptyString(models.currentModelId) ? models.currentModelId : null;
-  const currentValue =
-    (preferredValue && options.some((option) => option.value === preferredValue)
-      ? preferredValue
-      : options[0]?.value) ?? null;
-  if (!currentValue) {
-    return null;
-  }
-
-  return {
-    id: 'model',
-    name: 'Model',
-    type: 'select',
-    category: 'model',
-    currentValue,
-    options,
-  };
-}
-
-function resolveModeConfigOption(
-  modes: SessionModeState | null | undefined
-): SessionConfigOption | null {
-  if (!(modes && Array.isArray(modes.availableModes))) {
-    return null;
-  }
-
-  const options = modes.availableModes
-    .filter((entry) => isNonEmptyString(entry.id))
-    .map((entry) => ({
-      value: entry.id,
-      name: isNonEmptyString(entry.name) ? entry.name : entry.id,
-      ...(isNonEmptyString(entry.description) ? { description: entry.description } : {}),
-    }));
-  if (options.length === 0) {
-    return null;
-  }
-
-  const preferredValue = isNonEmptyString(modes.currentModeId) ? modes.currentModeId : null;
-  const currentValue =
-    (preferredValue && options.some((option) => option.value === preferredValue)
-      ? preferredValue
-      : options[0]?.value) ?? null;
-  if (!currentValue) {
-    return null;
-  }
-
-  return {
-    id: 'mode',
-    name: 'Mode',
-    type: 'select',
-    category: 'mode',
-    currentValue,
-    options,
-  };
-}
-
-function fallbackSessionConfigOptions(
-  sessionResult: Pick<SessionResultWithFallbackState, 'models' | 'modes'>
-): SessionConfigOption[] {
-  const fallbackOptions: SessionConfigOption[] = [];
-  const modelOption = resolveModelConfigOption(sessionResult.models);
-  if (modelOption) {
-    fallbackOptions.push(modelOption);
-  }
-  const modeOption = resolveModeConfigOption(sessionResult.modes);
-  if (modeOption) {
-    fallbackOptions.push(modeOption);
-  }
-  return fallbackOptions;
-}
-
-function requireSessionConfigOptions(
-  provider: string,
-  sessionSource: 'newSession' | 'loadSession' | 'setSessionConfigOption',
-  sessionResult: SessionResultWithFallbackState
-): SessionConfigOption[] {
-  let configOptions = Array.isArray(sessionResult.configOptions) ? sessionResult.configOptions : [];
-
-  if (sessionSource !== 'setSessionConfigOption') {
-    const fallbackOptions = fallbackSessionConfigOptions(sessionResult);
-    if (configOptions.length === 0 && fallbackOptions.length > 0) {
-      logger.warn('ACP session response missing configOptions; deriving from models/modes', {
-        provider,
-        sessionSource,
-      });
-      configOptions = fallbackOptions;
-    }
-  }
-
-  if (configOptions.length === 0) {
-    throw new Error(
-      `ACP ${provider} ${sessionSource} response did not include required configOptions`
-    );
-  }
-
-  const normalizedConfigOptions = normalizeSessionConfigOptions(provider, configOptions);
-  const missingCategories = REQUIRED_CONFIG_CATEGORIES.filter(
-    (category) => !normalizedConfigOptions.some((option) => option.category === category)
-  );
-  if (missingCategories.length > 0) {
-    throw new Error(
-      `ACP ${provider} ${sessionSource} response missing required config option categories: ${missingCategories.join(', ')}`
-    );
-  }
-
-  return normalizedConfigOptions;
 }
 
 export class AcpRuntimeManager {
