+++ src/client/routes/admin/index.ts
@@ -1,5 +1,6 @@
 export { type ApiUsageData, ApiUsageSection, type ApiUsageSectionProps } from './ApiUsageSection';
 export { IssueTrackingSection, ProjectIssueTrackingCard } from './IssueTrackingSection';
+export { PeriodicTasksSection } from './PeriodicTasksSection';
 export {
   type ProcessesData,
   ProcessesSection,
